+++ Sources/ArknightsClient/Features/Launcher/Settings/SettingsStrings.swift
@@ -133,10 +133,6 @@ enum SettingsStrings {
 		.settingsInstallationCanaryFeatures
 	static let canaryFeaturesDetail = LocalizedStringResource.Settings
 		.settingsInstallationCanaryFeaturesDetail
-	static let followsDefaultAudioOutput = LocalizedStringResource.Settings
-		.settingsInstallationFollowsDefaultAudioOutput
-	static let followsDefaultAudioOutputDetail = LocalizedStringResource.Settings
-		.settingsInstallationFollowsDefaultAudioOutputDetail
 	static let frameLatency = LocalizedStringResource.Settings.settingsInstallationFrameLatency
 	static let frameLatencyDetail = LocalizedStringResource.Settings
 		.settingsInstallationFrameLatencyDetail